First commit - code does not come from thin air!
This work is ported from initial development here: https://github.com/plusjade/ruhoh . I'm transfering the main ruby stuff into this gem. So far data processing works and I'm stubbing out a ruby based 'previewer'. The previewer's job is process full-previews of a user's pages/posts. Then any basic web-server can provide those views. It should not be the job of the gem to set up the web-server, that's what rack is for, so a config.ru will be provided in the main repo <--- this is important for cross-langauage compatability =)