Implement the Asset manager
Assets, defined as stylesheets,scripts, and media, are now handled collectively via the asset parser/manager. The parsing module 'finds and collects' assets from the current theme and all registered widgets.

Theme directory structure has been heavily re-implemented to allow for more intuitive organization of assets. Thanks to @filtercake for providing the inspiration.

Widgets may register script-based assets for inclusion into the page. Widgets cannot register stylesheets because all styling should be done by the theme. The theme can provide stylesheets on a per-widget basis which will automatically be detected and loaded if present.

To be clear the asset manager merely collects pointers to assets registered by the theme and widgets. Assets may be registered and grouped on a per-layout basis for a finer level of control. In development all assets are provided via {{ assets }} which should go in the head. As a TODO, when in production, the compiler should concatenate, minify and other-wise intelligently post-process the groups of assets.