The change of these files allow custom root url
The user can set production_url and dev_url variables in config.yml, if
they are not present, both are default to '/'.  Another variable env
determines whether production_url or dev_url is used as root url.  If
env == 'production', production_url is used, or else dev_url or simply
'/'.  The production_url can be a subdirectory on your www root
directory, or a virtual host, etc.