Allow the user to specify a custom scaffold url and support using mercurial(hg).
The user can specify the scaffold url on the command line like so...
    ruhoh new <blog name> <scaffold url>

If the user wants to use mercurial (hg) instead of git, they can simply add the
option `--hg` now.  The command will look like this...
    ruhoh --hg new <blog name>

The default scaffold is a git repository, so a mercurial user (like myself) will
either have to install the hg-git plugin for mercurial, or specify a custom
scaffold that is kept in a mercurial repository.