Fix links not properly rendering in the summary.
This change will search the rest of the content of your post and add any link
destinations to your summary before it gets rendered.  This is especially useful
for people who like to list their link destinations at the bottom of a page or
section, like so...

    [link1]: http://asdf.com/
    [link2]: http://google.com/

...and so forth.  Originally, link destinations in this format would be cut off
from the summary if they're past line 30 (by default), so any links using these
destinations in the summary weren't getting rendered properly.