Cache the config file
This commit caches the config file the first time it is read. It also
adds a special case to the resource watcher to reload the config on
change. While a separate config resource could be added, this seems like
overkill.

This halves the compile time on my system and should help fix
ruhoh/ruhoh.rb#166.