Add 'compile_as_root' configuration setting.
Ensures site is compiled from the root path even when base_path is set.
Solves the problem when a webhost such as GitHub hosts your site from a sub-directory but the folder you upload your website contents to is treated as the website "root".

Example: http://plusjade.github.io/cool-project

URLS on this website should have base_path : "/cool-project/" but the web-site directory structure should NOT be contained in a "cool-project" directory.