Huge refactor: Use Silly file-system query engine.
Silly (https://github.com/ruhoh/silly) abstracts the core data-modeling logic used to model files and their metadata attributes (top metadata). It also implements the cascade. Finally it provides the query interface for finding all files relative to the cascade and also using constraints based on metadata.

Silly is now the engine that powers ruhoh but to get to this point, much of ruhoh had to be blown up. Now ruhoh's job is to provide a more opionated 'web-site' user interface. Namely user configuration files, data conversion (markdown) and templating langauges (mustache) and layout support. Additionally the web-based previewer, compiler, and asset-pipeline etc.

Sorry for the ultra massive commit but all of this was a proof-of-concept until finally it was working.